# src/openapi_server/models/order.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401
from enum import Enum

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401


class OrderStatus(str, Enum):
    placed = 'placed'
    approved = 'approved'
    delivered = 'delivered'


class Order(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    Order - a model defined in OpenAPI

        id: The id of this Order [Optional].
        pet_id: The pet_id of this Order [Optional].
        quantity: The quantity of this Order [Optional].
        ship_date: The ship_date of this Order [Optional].
        status: The status of this Order [Optional].
        complete: The complete of this Order [Optional].
    """

    id: Optional[int] = Field(alias="id", default=None, example=10)
    pet_id: Optional[int] = Field(alias="petId", default=None, example=198772)
    quantity: Optional[int] = Field(alias="quantity", default=None, example=7)
    ship_date: Optional[datetime] = Field(alias="shipDate", default=None)
    status: Optional[OrderStatus] = Field(alias="status", default=None, description="Order Status", example="approved")
    complete: Optional[bool] = Field(alias="complete", default=None)

Order.update_forward_refs()