# src/openapi_server/models/pet.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401
from enum import Enum

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401
from openapi_server.models.category import Category
from openapi_server.models.tag import Tag


class PetStoreStatus(str, Enum):
    available = "available"
    pending = "pending"
    sold = "sold"


class Pet(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    Pet - a model defined in OpenAPI

        id: The id of this Pet [Optional].
        name: The name of this Pet.
        category: The category of this Pet [Optional].
        photo_urls: The photo_urls of this Pet.
        tags: The tags of this Pet [Optional].
        status: The status of this Pet [Optional].
    """

    id: Optional[int] = Field(alias="id", default=None, example=10)
    name: str = Field(alias="name", example="doggie")
    category: Optional[Category] = Field(alias="category", default=None)
    photo_urls: List[str] = Field(alias="photoUrls")
    tags: Optional[List[Tag]] = Field(alias="tags", default=None)
    status: Optional[PetStoreStatus] = Field(alias="status", default=None, description="pet status in the store")

Pet.update_forward_refs()